Cleanup deps, move glob_filter to own file
